feat: add stablePatchmarks option (#72)
Solves https://github.com/jest-community/snapshot-diff/issues/69.

I'm identifying patch marks with a simple regex that looks for a line starting with `@`. This seems safe since `jest-diff` appears to add whitespace at the start of all lines. Therefore, even a line of content that begins with `@` will end up padded with whitespace in the diff itself.

And rather than simply stripping out the patchmarks, I'm replacing them with `-------------`. Since the point of this is to avoid broken diffs, this solves that issue while still giving some indication that there are hidden lines in between those parts of the diff. Perhaps there's a better replacement indicator than that. We could certainly preserve the patch mark `@@` notation like `@@ ------------- @@` if we think that helps communicate the meaning of the new mark.